SOURCE CODE/CLIENT/requirements.txt
pillow
pytesseract
opencv-python
pyautogui